Static-analysis and include hygiene in CI (#19)
Adds whole-program and include checks to the lint pipeline and locks in
a few compiler guarantees. No behavior change to the decoder on valid
input.

- **cppcheck (script/cppcheck.sh + CI job):** whole-program pass; its
unusedFunction check sees across translation units where per-TU
clang-tidy can't. It found no existing issues.
- **Warning flags:** `-Wmissing-prototypes`/`-Wmissing-declarations`
(any non-header-declared function must be static) and
`-Wold-style-cast`; both already clean, the flags lock the state in.
`decode()` marked `[[nodiscard]]` (GNU `warn_unused_result` fallback at
C++11) so ignoring the sole error channel warns.
- **buf_ fill bound:** explicit `buf_len_ < sizeof(buf_)` bound at all
three accumulation sites, making the header-validated invariant visible
to clang-analyzer's ArrayBound check. No change on valid state; degrades
to `NEED_MORE_DATA` rather than overflowing if state were ever
corrupted.
- **include-what-you-use (script/check-includes.sh + CI job):**
clang-include-cleaner over every source/header in a host pass and an ESP
pass (`-DESP_PLATFORM` + stub ESP-IDF headers) so guarded ESP32 branches
are checked without a cross toolchain. Surfaced two test files using
fixed-width int types without a direct `<cstdint>`.

+++ tests/test_wav_decoder.cpp
@@ -15,6 +15,7 @@
 #include "micro_wav/wav_decoder.h"
 #include "wav_test_data.h"
 
+#include <cstdint>
 #include <cstdio>
 #include <cstdlib>
 #include <cstring>
@@ -558,7 +559,7 @@ static bool test_streaming_byte_by_byte() {
         WAVDecoder decoder;
         WAVDecoderResult result = decode_header_chunked(decoder, tc.data, tc.len, 1);
         if (result != WAV_DECODER_HEADER_READY) {
-            fprintf(stderr, "  FAIL: streaming byte-by-byte %s: bad result %d\n", tc.name, result);
+            fprintf(stderr, "  FAIL: streaming byte-by-byte %s: bad result %d\n", tc.name, static_cast<int>(result));
             return false;
         }
         if (decoder.get_audio_format() != tc.expected_format ||
@@ -578,7 +579,7 @@ static bool test_streaming_two_bytes() {
         WAVDecoder decoder;
         WAVDecoderResult result = decode_header_chunked(decoder, tc.data, tc.len, 2);
         if (result != WAV_DECODER_HEADER_READY) {
-            fprintf(stderr, "  FAIL: streaming 2-byte %s: bad result %d\n", tc.name, result);
+            fprintf(stderr, "  FAIL: streaming 2-byte %s: bad result %d\n", tc.name, static_cast<int>(result));
             return false;
         }
         if (decoder.get_audio_format() != tc.expected_format ||
@@ -598,7 +599,7 @@ static bool test_streaming_three_bytes() {
         WAVDecoder decoder;
         WAVDecoderResult result = decode_header_chunked(decoder, tc.data, tc.len, 3);
         if (result != WAV_DECODER_HEADER_READY) {
-            fprintf(stderr, "  FAIL: streaming 3-byte %s: bad result %d\n", tc.name, result);
+            fprintf(stderr, "  FAIL: streaming 3-byte %s: bad result %d\n", tc.name, static_cast<int>(result));
             return false;
         }
         if (decoder.get_audio_format() != tc.expected_format ||
@@ -618,7 +619,7 @@ static bool test_streaming_five_bytes() {
         WAVDecoder decoder;
         WAVDecoderResult result = decode_header_chunked(decoder, tc.data, tc.len, 5);
         if (result != WAV_DECODER_HEADER_READY) {
-            fprintf(stderr, "  FAIL: streaming 5-byte %s: bad result %d\n", tc.name, result);
+            fprintf(stderr, "  FAIL: streaming 5-byte %s: bad result %d\n", tc.name, static_cast<int>(result));
             return false;
         }
         if (decoder.get_audio_format() != tc.expected_format ||
